feat: Pretendard 폰트 설정 추가 (#7)
- localFont를 사용해 Pretendard 폰트 설정
