+++ app/(home)/layout.tsx
@@ -1,20 +1,9 @@
-import type { Metadata } from 'next'
-
-export const metadata: Metadata = {
-  title: 'InvestMetic',
-  description: '',
-}
-
-const RootLayout = ({
-  children,
-}: Readonly<{
+interface Props {
   children: React.ReactNode
-}>) => {
-  return (
-    <html lang="ko">
-      <body>{children}</body>
-    </html>
-  )
 }
 
-export default RootLayout
+const HomeLayout = ({ children }: Props) => {
+  return <>{children}</>
+}
+
+export default HomeLayout
